# Conflicts:
#	.gitlab-ci.yml

+++ pkg/aggregator/mocksender/asserts.go
@@ -32,7 +32,7 @@ func (m *MockSender) AssertMetric(t *testing.T, method string, metric string, va
 
 // AssertHistogramBucket allows to assert a histogram bucket was emitted with given parameters.
 // Additional tags over the ones specified don't make it fail
-func (m *MockSender) AssertHistogramBucket(t *testing.T, method string, metric string, value int, lowerBound float64, upperBound float64, monotonic bool, hostname string, tags []string) bool {
+func (m *MockSender) AssertHistogramBucket(t *testing.T, method string, metric string, value int64, lowerBound float64, upperBound float64, monotonic bool, hostname string, tags []string) bool {
 	return m.Mock.AssertCalled(t, method, metric, value, lowerBound, upperBound, monotonic, hostname, tags)
 }
