+++ rtloader/test/aggregator/aggregator.go
@@ -17,10 +17,10 @@ import (
 #include "rtloader_mem.h"
 #include "datadog_agent_rtloader.h"
 
-extern void submitMetric(char *, metric_type_t, char *, float, char **, char *);
+extern void submitMetric(char *, metric_type_t, char *, double, char **, char *);
 extern void submitServiceCheck(char *, char *, int, char **, char *, char *);
 extern void submitEvent(char*, event_t*);
-extern void submitHistogramBucket(char *, char *, int, float, float, int, char *, char **);
+extern void submitHistogramBucket(char *, char *, long long, float, float, int, char *, char **);
 
 static void initAggregatorTests(rtloader_t *rtloader) {
    set_submit_metric_cb(rtloader, submitMetric);
@@ -151,7 +151,7 @@ func charArrayToSlice(array **C.char) (res []string) {
 }
 
 //export submitMetric
-func submitMetric(id *C.char, mt C.metric_type_t, mname *C.char, val C.float, t **C.char, hname *C.char) {
+func submitMetric(id *C.char, mt C.metric_type_t, mname *C.char, val C.double, t **C.char, hname *C.char) {
 	checkID = C.GoString(id)
 	metricType = int(mt)
 	name = C.GoString(mname)
@@ -211,7 +211,7 @@ func submitEvent(id *C.char, ev *C.event_t) {
 }
 
 //export submitHistogramBucket
-func submitHistogramBucket(id *C.char, cMetricName *C.char, cVal C.int, cLowerBound C.float, cUpperBound C.float, cMonotonic C.int, cHostname *C.char, t **C.char) {
+func submitHistogramBucket(id *C.char, cMetricName *C.char, cVal C.longlong, cLowerBound C.float, cUpperBound C.float, cMonotonic C.int, cHostname *C.char, t **C.char) {
 	checkID = C.GoString(id)
 	name = C.GoString(cMetricName)
 	intValue = int(cVal)
